Adds broken redux detection code
I'm trying to make it so that vim can detect redux files, but I have yet
to find the best place to put the code. I want to have a system where
the new filetype is an augmentation (i.e. the file is still javascript,
but it is also redux - therefore should use the redux.javascript to have
access to redux snippets etc), but it seems like the way that Vim is set
up it wants it to be either javascript OR redux. Or at least, it wants
me to explicitly set the filetype to "redux.javascript" rather than
allowing me to augment the file type. I want to be able to augment it in
case the filetype is javascript.jsx - I don't want to lose the jsx part
by explicitly setting the filetype to redux.javascript.