Consolidate vim and mutt config files
It seems v hard to remove the .vim and .mutt directories from my $HOME
dir. But I can at least make it so that they are the only foot print
that vim and mutt leave in the home dir. This is what this commit does.
Might require some re linking on the work laptop...